Fix time zone bugs
The execution date is send to server as UTC and get from server in UTC format too.
The execution date is displayed according to the current time zone of the client.
The month time range use time zone form X-Time-Zone header parameter.
It was created a new version of the REST API in summary-service. X-API-Version 1.1